Move parsing and config tests; reattribute download comment
