+++ crates/triplets-hf-source/src/config.rs
@@ -173,6 +173,8 @@ impl HuggingFaceRowsConfig {
 mod tests {
     use super::*;
     use crate::constants::HF_PARQUET_DEFAULT_ENDPOINT;
+    use crate::source_core::HuggingFaceRowSource;
+    use crate::test_utils::test_config;
     use tempfile::tempdir;
 
     #[test]
@@ -185,5 +187,17 @@ mod tests {
         let c =
             HuggingFaceRowsConfig::new("ep_test", "org/dataset", "default", "train", dir.path());
         assert_eq!(c.parquet_endpoint, HF_PARQUET_DEFAULT_ENDPOINT);
+    }
+
+    #[test]
+    fn normalized_shard_extensions_trims_dots_and_lowercases() {
+        let dir = tempdir().unwrap();
+        let mut config = test_config(dir.path().to_path_buf());
+        config.shard_extensions = vec![".PARQUET".into(), " ndjson ".into()];
+        let normalized = HuggingFaceRowSource::normalized_shard_extensions(&config);
+        assert_eq!(
+            normalized,
+            vec!["parquet".to_string(), "ndjson".to_string()]
+        );
     }
 }
